🎨 Separate roomManagement and spaceManagement
